+++ SRC/Private/Abstractions/Interfaces/Strategies/ITypeResolution.cs
@@ -4,39 +4,18 @@
 * Author: Denes Solti                                                           *
 ********************************************************************************/
 using System;
-using System.Reflection;
 using System.Threading;
 
 namespace Solti.Utils.Proxy.Internals
 {
     /// <summary>
-    /// Represents the method how the generated <see cref="Type"/> will be resolved.
+    /// Represents the method how the proxy <see cref="Type"/> will be resolved.
     /// </summary>
     internal interface ITypeResolution
     {
         /// <summary>
-        /// The related <see cref="TypeGenerator{TDescendant}"/> descendant.
+        /// Tries to resolves the generated <see cref="Type"/>.
         /// </summary>
-        Type GeneratorType { get; }
-
-        /// <summary>
-        /// The name of the <see cref="Assembly"/> that contains the generated <see cref="Type"/>.
-        /// </summary>
-        string ContainingAssembly { get; }
-
-        /// <summary>
-        /// The fully qualified name of the generated <see cref="Type"/>.
-        /// </summary>
-        string ClassName { get; }
-
-        /// <summary>
-        /// Returns true if the strategy should be used.
-        /// </summary>
-        bool ShouldUse { get; }
-
-        /// <summary>
-        /// Resolves the generated <see cref="Type"/>.
-        /// </summary>
-        Type Resolve(CancellationToken cancellation = default);
+        Type? TryResolve(CancellationToken cancellation = default);
     }
 }