+++ SRC/Private/SyntaxFactories/ProxySyntaxFactory.EventInterceptorFactory.cs
@@ -3,6 +3,7 @@
 *                                                                               *
 * Author: Denes Solti                                                           *
 ********************************************************************************/
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -17,39 +18,34 @@ namespace Solti.Utils.Proxy.Internals
     internal partial class ProxySyntaxFactory
     {
         /// <summary>
-        /// event EventType IInterface.Event                                            <br/>
-        /// {                                                                           <br/>
-        ///     add                                                                     <br/>
-        ///     {                                                                       <br/>
-        ///         object[] args = new object[] {value};                               <br/>
-        ///         InvokeTarget = () =>                                                <br/>
-        ///         {                                                                   <br/>
-        ///             EventType cb_value = (EventType) args[a];                       <br/>
-        ///             Target.Event += cb_value;                                       <br/>
-        ///             return null;                                                    <br/>
-        ///         };                                                                  <br/>
-        ///         EventInfo evt = ResolveEvent(InvokeTarget);                         <br/>
-        ///         Invoke(evt.AddMethod, args, evt);                                   <br/>
-        ///     }                                                                       <br/>
-        ///     remove                                                                  <br/>
-        ///     {                                                                       <br/>
-        ///         object[] args = new object[] {value};                               <br/>
-        ///         InvokeTarget = () =>                                                <br/>
-        ///         {                                                                   <br/>
-        ///             EventType cb_value = (EventType) args[a];                       <br/>
-        ///             Target.Event -= cb_value;                                       <br/>
-        ///             return null;                                                    <br/>
-        ///         };                                                                  <br/>
-        ///         EventInfo evt = ResolveEvent(InvokeTarget);                         <br/>
-        ///         Invoke(evt.RemoveMethod, args, evt);                                <br/>
-        ///     }                                                                       <br/>
+        /// event EventType IInterface.Event                                                  <br/>
+        /// {                                                                                 <br/>
+        ///     add                                                                           <br/>
+        ///     {                                                                             <br/>
+        ///         object[] args = new object[] {value};                                     <br/>
+        ///         Func[object] invokeTarget = () =>                                         <br/>
+        ///         {                                                                         <br/>
+        ///             EventType cb_value = (EventType) args[a];                             <br/>
+        ///             Target.Event += cb_value;                                             <br/>
+        ///             return null;                                                          <br/>
+        ///         };                                                                        <br/>
+        ///         Invoke(new InvocationContext(args, invokeTarget, MemberTypes.Event));     <br/>
+        ///     }                                                                             <br/>
+        ///     remove                                                                        <br/>
+        ///     {                                                                             <br/>
+        ///         object[] args = new object[] {value};                                     <br/>
+        ///         Func[object] invokeTarget = () =>                                         <br/>
+        ///         {                                                                         <br/>
+        ///             EventType cb_value = (EventType) args[a];                             <br/>
+        ///             Target.Event -= cb_value;                                             <br/>
+        ///             return null;                                                          <br/>
+        ///         };                                                                        <br/>
+        ///         Invoke(new InvocationContext(args, invokeTarget, MemberTypes.Event));     <br/>
+        ///     }                                                                             <br/>
         /// }
         /// </summary>
         internal sealed class EventInterceptorFactory : ProxyMemberSyntaxFactory
         {
-            private readonly IMethodInfo
-                RESOLVE_EVENT;
-
             private IEnumerable<StatementSyntax> Build(IEventInfo member, bool add) 
             {
                 IMethodInfo? targetMethod = add ? member.AddMethod : member.RemoveMethod;
@@ -59,8 +55,9 @@ namespace Solti.Utils.Proxy.Internals
                 LocalDeclarationStatementSyntax argsArray = CreateArgumentsArray(targetMethod);
                 yield return argsArray;
 
-                yield return AssignCallback
+                LocalDeclarationStatementSyntax invokeTarget = DeclareLocal<Func<object>>
                 (
+                    nameof(invokeTarget),
                     DeclareCallback
                     (
                         argsArray,
@@ -81,19 +78,7 @@ namespace Solti.Utils.Proxy.Internals
                         }
                     )
                 );
-
-                LocalDeclarationStatementSyntax evt = DeclareLocal<EventInfo>(nameof(evt), InvokeMethod
-                (
-                    RESOLVE_EVENT,
-                    target: null,
-                    castTargetTo: null,
-                    Argument
-                    (
-                        expression: PropertyAccess(INVOKE_TARGET, null, null)
-                    )
-                ));
-
-                yield return evt;
+                yield return invokeTarget;
 
                 yield return ExpressionStatement
                 (
@@ -102,32 +87,21 @@ namespace Solti.Utils.Proxy.Internals
                         INVOKE,
                         target: null,
                         castTargetTo: null,
-                        arguments: new ExpressionSyntax[]
-                        {
-                            SimpleMemberAccess // evt.[Add|Remove]Method
+                        Argument
+                        (
+                            CreateObject<InvocationContext>
                             (
-                                ToIdentifierName(evt),
-                                add ? nameof(EventInfo.AddMethod) : nameof(EventInfo.RemoveMethod)
-                            ),
-                            ToIdentifierName(argsArray), // args
-                            ToIdentifierName(evt) // evt
-                        }.Select(Argument).ToArray()
+                                ToArgument(argsArray),
+                                ToArgument(invokeTarget),
+                                Argument(EnumAccess(MemberTypes.Event))
+                            )
+                        )
                     )
                 );
             }
 
             public EventInterceptorFactory(IProxyContext context) : base(context)
             {
-                RESOLVE_EVENT = Context.InterceptorType.Methods.Single
-                (
-                    met => met.SignatureEquals
-                    (
-                        MetadataMethodInfo.CreateFrom
-                        (
-                            (MethodInfo) MemberInfoExtensions.ExtractFrom(() => InterfaceInterceptor<object>.ResolveEvent(default!))
-                        )
-                    )
-                );
             }
 
             protected override IEnumerable<MemberDeclarationSyntax> BuildMembers(CancellationToken cancellation) => Context