+++ SRC/Private/SyntaxFactories/ProxySyntaxFactory.PropertyInterceptorFactory.cs
@@ -19,37 +19,32 @@ namespace Solti.Utils.Proxy.Internals
     internal partial class ProxySyntaxFactory
     {
         /// <summary>
-        /// TResult IInterface.Prop                                                          <br/>
-        /// {                                                                                <br/>
-        ///     get                                                                          <br/>
-        ///     {                                                                            <br/>
-        ///         object[] args = new object[] { };                                        <br/>
-        ///         InvokeTarget = () =>                                                     <br/>
-        ///         {                                                                        <br/>
-        ///             return Target.Prop;                                                  <br/>
-        ///         };                                                                       <br/>
-        ///         PropertyInfo prop = ResolveProperty(InvokeTarget);                       <br/>
-        ///         return (TResult) Invoke(prop.GetMethod, args, prop);                     <br/>
-        ///     }                                                                            <br/>
-        ///     set                                                                          <br/>
-        ///     {                                                                            <br/>
-        ///         object[] args = new object[] {value};                                    <br/>
-        ///         InvokeTarget = () =>                                                     <br/>
-        ///         {                                                                        <br/>
-        ///           TResult cb_value = (TResult) args[0];                                  <br/>
-        ///           Target.Prop = cb_value;                                                <br/>
-        ///           return null;                                                           <br/>
-        ///         };                                                                       <br/>
-        ///         PropertyInfo prop = ResolveProperty(InvokeTarget);                       <br/>
-        ///         Invoke(prop.SetMethod, args, prop);                                      <br/>
-        ///     }                                                                            <br/>
+        /// TResult IInterface.Prop                                                                           <br/>
+        /// {                                                                                                 <br/>
+        ///     get                                                                                           <br/>
+        ///     {                                                                                             <br/>
+        ///         object[] args = new object[] { };                                                         <br/>
+        ///         Func[object] invokeTarget = () =>                                                         <br/>
+        ///         {                                                                                         <br/>
+        ///             return Target.Prop;                                                                   <br/>
+        ///         };                                                                                        <br/>
+        ///         return (TResult) Invoke(new InvocationContext(args, invokeTarget, MemberTypes.Property)); <br/>
+        ///     }                                                                                             <br/>
+        ///     set                                                                                           <br/>
+        ///     {                                                                                             <br/>
+        ///         object[] args = new object[] {value};                                                     <br/>
+        ///         Func[object] invokeTarget = () =>                                                         <br/>
+        ///         {                                                                                         <br/>
+        ///           TResult cb_value = (TResult) args[0];                                                   <br/>
+        ///           Target.Prop = cb_value;                                                                 <br/>
+        ///           return null;                                                                            <br/>
+        ///         };                                                                                        <br/>
+        ///         Invoke(new InvocationContext(args, invokeTarget, MemberTypes.Property));                  <br/>
+        ///     }                                                                                             <br/>
         /// }
         /// </summary>
         internal sealed class PropertyInterceptorFactory : ProxyMemberSyntaxFactory
         {
-            private readonly IMethodInfo
-                RESOLVE_PROPERTY;
-
             private IEnumerable<StatementSyntax> BuildGet(IPropertyInfo property) 
             {
                 if (property.GetMethod is null) yield break;
@@ -57,8 +52,9 @@ namespace Solti.Utils.Proxy.Internals
                 LocalDeclarationStatementSyntax argsArray = CreateArgumentsArray(property.GetMethod);
                 yield return argsArray;
 
-                yield return AssignCallback
+                LocalDeclarationStatementSyntax invokeTarget = DeclareLocal<Func<object>>
                 (
+                    nameof(invokeTarget),
                     DeclareCallback
                     (
                         argsArray,
@@ -78,19 +74,7 @@ namespace Solti.Utils.Proxy.Internals
                         )
                     )
                 );
-
-                LocalDeclarationStatementSyntax prop = DeclareLocal<PropertyInfo>(EnsureUnused(nameof(prop), property.GetMethod), InvokeMethod
-                (
-                    RESOLVE_PROPERTY,
-                    target: null,
-                    castTargetTo: null,
-                    Argument
-                    (
-                        expression: PropertyAccess(INVOKE_TARGET, null, null)
-                    )
-                ));
-
-                yield return prop;
+                yield return invokeTarget;
 
                 yield return ReturnResult
                 (
@@ -100,16 +84,15 @@ namespace Solti.Utils.Proxy.Internals
                         INVOKE,
                         target: null,
                         castTargetTo: null,
-                        arguments: new ExpressionSyntax[]
-                        {
-                            SimpleMemberAccess // prop.GetMethod
+                        Argument
+                        (
+                            CreateObject<InvocationContext>
                             (
-                                ToIdentifierName(prop),  
-                                nameof(PropertyInfo.GetMethod)
-                            ),
-                            ToIdentifierName(argsArray), // new object[0] | new object[] {index1, index2, ...}       
-                            ToIdentifierName(prop) // prop
-                        }.Select(Argument).ToArray()
+                                ToArgument(argsArray),
+                                ToArgument(invokeTarget),
+                                Argument(EnumAccess(MemberTypes.Property))
+                            )
+                        )
                     )
                 );
             }
@@ -121,8 +104,9 @@ namespace Solti.Utils.Proxy.Internals
                 LocalDeclarationStatementSyntax argsArray = CreateArgumentsArray(property.SetMethod);
                 yield return argsArray;
 
-                yield return AssignCallback
+                LocalDeclarationStatementSyntax invokeTarget = DeclareLocal<Func<object>>
                 (
+                    nameof(invokeTarget),
                     DeclareCallback
                     (
                         argsArray,
@@ -137,10 +121,10 @@ namespace Solti.Utils.Proxy.Internals
                                     (
                                         kind: SyntaxKind.SimpleAssignmentExpression,
                                         left: PropertyAccess(property, MemberAccess(null, TARGET), null, locals
-#if NETSTANDARD2_0
-                                            .Take(locals.Count - 1)
+#if NETSTANDARD2_1_OR_GREATER
+                                            .SkipLast(1)                                          
 #else
-                                            .SkipLast(1)
+                                            .Take(locals.Count - 1)
 #endif
                                             .Select(ToArgument)),
                                         right: ToIdentifierName(locals[locals.Count - 1])
@@ -154,19 +138,7 @@ namespace Solti.Utils.Proxy.Internals
                         }
                     )
                 );
-
-                LocalDeclarationStatementSyntax prop = DeclareLocal<PropertyInfo>(EnsureUnused(nameof(prop), property.SetMethod), InvokeMethod
-                (
-                    RESOLVE_PROPERTY,
-                    target: null,
-                    castTargetTo: null,
-                    Argument
-                    (
-                        expression: PropertyAccess(INVOKE_TARGET, null, null)
-                    )
-                ));
-
-                yield return prop;
+                yield return invokeTarget;
 
                 yield return ExpressionStatement
                 (
@@ -175,16 +147,15 @@ namespace Solti.Utils.Proxy.Internals
                         INVOKE,
                         target: null,
                         castTargetTo: null,
-                        arguments: new ExpressionSyntax[]
-                        {
-                            SimpleMemberAccess // prop.SetMethod
+                        Argument
+                        (
+                            CreateObject<InvocationContext>
                             (
-                                ToIdentifierName(prop),
-                                nameof(PropertyInfo.SetMethod)
-                            ),
-                            ToIdentifierName(argsArray), //  new object[] {value} | new object[] {index1, index2, ..., value}
-                            ToIdentifierName(prop) // prop
-                        }.Select(Argument).ToArray()
+                                ToArgument(argsArray),
+                                ToArgument(invokeTarget),
+                                Argument(EnumAccess(MemberTypes.Property))
+                            )
+                        )
                     )
                 );
             }
@@ -226,16 +197,6 @@ namespace Solti.Utils.Proxy.Internals
 
             public PropertyInterceptorFactory(IProxyContext context) : base(context) 
             {
-                RESOLVE_PROPERTY = Context.InterceptorType.Methods.Single
-                (
-                    met => met.SignatureEquals
-                    (
-                        MetadataMethodInfo.CreateFrom
-                        (
-                            (MethodInfo) MemberInfoExtensions.ExtractFrom(() => InterfaceInterceptor<object>.ResolveProperty(default!))
-                        )
-                    )
-                );
             }
         }
     }